+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Upgrade to wombat 3.8.11 (#256)
 
+### Fixed
+
+- JS rewriting abusively rewrite import function (#255)
+
 ## [5.1.1] - 2025-02-17
 
 ### Changed